src/test/javascript/spec/app/entities/vendor-wh-courier-mapping/vendor-wh-courier-mapping.service.spec.ts
/* tslint:disable max-line-length */
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { VendorWHCourierMappingService } from 'app/entities/vendor-wh-courier-mapping/vendor-wh-courier-mapping.service';
import { VendorWHCourierMapping } from 'app/shared/model/vendor-wh-courier-mapping.model';
import { SERVER_API_URL } from 'app/app.constants';

describe('Service Tests', () => {
    describe('VendorWHCourierMapping Service', () => {
        let injector: TestBed;
        let service: VendorWHCourierMappingService;
        let httpMock: HttpTestingController;

        beforeEach(() => {
            TestBed.configureTestingModule({
                imports: [HttpClientTestingModule]
            });
            injector = getTestBed();
            service = injector.get(VendorWHCourierMappingService);
            httpMock = injector.get(HttpTestingController);
        });

        describe('Service methods', () => {
            it('should call correct URL', () => {
                service.find(123).subscribe(() => {});

                const req = httpMock.expectOne({ method: 'GET' });

                const resourceUrl = SERVER_API_URL + 'api/vendor-wh-courier-mappings';
                expect(req.request.url).toEqual(resourceUrl + '/' + 123);
            });

            it('should create a VendorWHCourierMapping', () => {
                service.create(new VendorWHCourierMapping(null)).subscribe(received => {
                    expect(received.body.id).toEqual(null);
                });

                const req = httpMock.expectOne({ method: 'POST' });
                req.flush({ id: null });
            });

            it('should update a VendorWHCourierMapping', () => {
                service.update(new VendorWHCourierMapping(123)).subscribe(received => {
                    expect(received.body.id).toEqual(123);
                });

                const req = httpMock.expectOne({ method: 'PUT' });
                req.flush({ id: 123 });
            });

            it('should return a VendorWHCourierMapping', () => {
                service.find(123).subscribe(received => {
                    expect(received.body.id).toEqual(123);
                });

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush({ id: 123 });
            });

            it('should return a list of VendorWHCourierMapping', () => {
                service.query(null).subscribe(received => {
                    expect(received.body[0].id).toEqual(123);
                });

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush([new VendorWHCourierMapping(123)]);
            });

            it('should delete a VendorWHCourierMapping', () => {
                service.delete(123).subscribe(received => {
                    expect(received.url).toContain('/' + 123);
                });

                const req = httpMock.expectOne({ method: 'DELETE' });
                req.flush(null);
            });

            it('should propagate not found response', () => {
                service.find(123).subscribe(null, (_error: any) => {
                    expect(_error.status).toEqual(404);
                });

                const req = httpMock.expectOne({ method: 'GET' });
                req.flush('Invalid request parameters', {
                    status: 404,
                    statusText: 'Bad Request'
                });
            });
        });

        afterEach(() => {
            httpMock.verify();
        });
    });
});
